Slip swap form into sub components
